requirements.txt
django==3.1
django-extensions==3.1.1
factory-boy==2.12.0
wagtail==2.12
psycopg2-binary==2.8.6
djangorestframework==3.12.2
wagtail-factories==2.0.1